added the ability to change the users and fixed a input validation bug, also adjusted the listAll() function to show name data, added description for create() function
first, i added a description of what the create function does.  added
code to the add all function that allows the admin(s) to see what names
of the users are if there is one.  fixed a bug in the create name
function that prevented users with different types of names from adding
them properly.  created the changeName() function, this allows users to
update their name that we have stored.